+++ DelvUI/changelog.md
@@ -1,9 +1,10 @@
 # 0.6.3.0
 Features:
-- Added setting to status effects lists to disable mouse interactions.
+- Added setting to Status Effects Lists to disable mouse interactions.
+- Added target border thickness setting to the Enemy List.
 
 Fixes:
-- Fixed party frames border color not working.
+- Fixed Party Frames border color not working.
 - Fixed party members health label not visible on preview mode.
 
 # 0.6.2.1